+++ requirments.txt
@@ -1,0 +1,3 @@
+soundfile
+matplotlib.pyplot
+ibrosa, librosa.display
