+++ src/process.py
@@ -1,21 +1,6 @@
-#라이브러리 설치
-import numpy as np
-import librosa, librosa.display
 import matplotlib.pyplot as plt
-import soundfile as sf
+import librosa, librosa.display
 
-#----------------------TO DO----------------------------
-# 특징 추출 + 노이즈 제거 등
-#-------------------------------------------------------
-
-
-# wav 파일 불러오기
-file = "src/example.wav"
-
-# Librosa로 오디오 불러오기
-y, sr = librosa.load(file) 
-# y : 파형의 amplitude값, sr : sampling rate(초당 샘플 갯수) 기본적으로 22500
-# sr = 22500 은 1초당 22500개의 데이터를 샘플링하는 것.
 
 # Draw WavePlot
 def draw_waveplot(y, sr):
@@ -44,7 +29,6 @@ def draw_spectrogram(y, sr):
 
 
 # wav 파일 Edit(공백제거 --> 무음 구간 기준으로 나누기)
-# 우선 절반으로 나눠보기
 def wav_del_space(y, sr):
 
     segment = []
@@ -58,17 +42,3 @@ def wav_del_space(y, sr):
         segment.append(y[start:end]) # 무음이 아닌 구간을 segment 리스트에 추가
 
     return segment
-    
-# Sav edit FIlE
-def save(filename, segment, sr):
-    
-    for i, segment in enumerate(segment):
-        save_name = f"{filename}_{i}.wav"
-
-        sf.write(save_name, segment, sr)  # 편집된 오디오 파일 저장
-        print(f"Saved: {save_name}")
-
-
-# 실행
-seg = wav_del_space(y, sr)
-save("src/example_edited", seg, sr)
